+++ email_strings.py
@@ -1,0 +1,11 @@
+INVITE_LINK_EMAIL = """Subject: Invitation to register at %%organisation%%
+
+Hi!
+
+You've been invited to signup for an %%organisation%% account. Please go to the following url to signup: %%link%%
+
+This link is valid for %%days%% days. If you need a new link, please contact the person who invited you to this organisation.
+
+All the best,
+%%organisation%% 
+"""