+++ help_strings.py
@@ -12,3 +12,13 @@ HELP_USERS_CREATE = """Create one or more users.
 Takes one or more email address as parameters. Creates the users in the identity provider,
 marking their emails as verified. Then sends them an email with a password reset link.
 """
+
+HELP_USERS_INVITE = """Invite one or more users.
+
+Takes one or more email address as parameters. Creates a self-registration page for each user
+and sends the user a link to it.
+"""
+
+HELP_USERS_KEYCLOAK_DISABLED = "The users command is not configured on this instance, sorry."
+
+HELP_USERS_KEYCLOAK_SIGNUP_DISABLED = "The users invite command is not configured on this instance, sorry."